change pagination

+++ contests/models.py
@@ -517,7 +517,7 @@ class ModxDbimgMuz(models.Model):
 
 
 class Archive(BaseContest):
-    rating = models.IntegerField('Райтинг', blank=True, default=0)
+    rating = models.FloatField('Райтинг', blank=True, default=0)
     fio = models.CharField('Участник', max_length=700, blank=True, default='')
     content = RichTextField(verbose_name='Контент', blank=True, null=True)
     publish = models.BooleanField(verbose_name='Опубликовать', default=False)
